+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 
 ## [Unreleased]
 ### Added
+
+- Permissions.instances(): add `explicit` keyword argument
+- Permissions.instances(): add `ignore_grant_all` keyword argument
+- Permissions.check(): add `ignore_grant_all` keyword argument
+
 ### Fixed
 ### Changed
 ### Deprecated